Use fewer ints

+++ geojson.hpp
@@ -11,12 +11,12 @@
 
 struct parse_json_args {
 	json_pull *jp;
-	int layer;
+	size_t layer;
 	std::string *layername;
 
 	struct serialization_state *sst;
 
-	parse_json_args(json_pull *jp1, int layer1, std::string *layername1, struct serialization_state *sst1)
+	parse_json_args(json_pull *jp1, size_t layer1, std::string *layername1, struct serialization_state *sst1)
 	    : jp(jp1), layer(layer1), layername(layername1), sst(sst1) {
 	}
 };
@@ -24,7 +24,7 @@ struct parse_json_args {
 struct json_pull *json_begin_map(char *map, long len);
 void json_end_map(struct json_pull *jp);
 
-void parse_json(struct serialization_state *sst, json_pull *jp, int layer, std::string layername);
+void parse_json(struct serialization_state *sst, json_pull *jp, size_t layer, std::string layername);
 void *run_parse_json(void *v);
 
 #endif
